Support labeled histogram and percentile

+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/meter/function/BucketedValues.java
@@ -40,9 +40,9 @@ public class BucketedValues {
      * The element in the buckets represent the minimal value of this bucket, the max is defined by the next element.
      * Such as 0, 10, 50, 100 means buckets are [0, 10), [10, 50), [50, 100), [100, infinite+).
      *
-     * The {@link Integer#MIN_VALUE} could be the first bucket element to indicate there is no minimal value.
+     * The {@link Long#MIN_VALUE} could be the first bucket element to indicate there is no minimal value.
      */
-    private int[] buckets;
+    private long[] buckets;
     /**
      * {@link #buckets} and {@link #values} arrays should have the same length. The element in the values, represents
      * the amount in the same index bucket.
@@ -53,7 +53,7 @@ public class BucketedValues {
      * @param buckets Read {@link #buckets}
      * @param values  Read {@link #values}
      */
-    public BucketedValues(final int[] buckets, final long[] values) {
+    public BucketedValues(final long[] buckets, final long[] values) {
         if (buckets == null || values == null || buckets.length == 0 || values.length == 0) {
             throw new IllegalArgumentException("buckets and values can't be null.");
         }
@@ -69,13 +69,13 @@ public class BucketedValues {
      */
     public boolean isCompatible(DataTable dataset) {
         final List<String> sortedKeys = dataset.sortedKeys(new HeatMap.KeyComparator(true));
-        int[] existedBuckets = new int[sortedKeys.size()];
+        long[] existedBuckets = new long[sortedKeys.size()];
         for (int i = 0; i < sortedKeys.size(); i++) {
             final String key = sortedKeys.get(i);
             if (key.equals(Bucket.INFINITE_NEGATIVE)) {
-                existedBuckets[i] = Integer.MIN_VALUE;
+                existedBuckets[i] = Long.MIN_VALUE;
             } else {
-                existedBuckets[i] = Integer.parseInt(key);
+                existedBuckets[i] = Long.parseLong(key);
             }
         }
